feat: reindexing should be at the end of imports and refactored duplicate code
