+++ packages/vendure-scripts/src/assign-products-to-channel/assign-all-products-to-channel.ts
@@ -1,13 +1,6 @@
-import {
-  ID,
-  Injector,
-  Product,
-  RequestContext,
-  TransactionalConnection,
-} from '@vendure/core';
-import { assignTheseProductsToChannel } from './assign-these-products-to-channel';
-import { getProductsDeep } from './get-products-deep';
+import { ID, Injector, RequestContext } from '@vendure/core';
 import { IsNull } from 'typeorm';
+import { assignProductsInBatch } from './batch-assign-products';
 
 export async function assignAllProductsToChannel(
   sourceChannelId: ID,
@@ -16,30 +9,14 @@ export async function assignAllProductsToChannel(
   ctx: RequestContext,
   batch: number = 10
 ): Promise<void> {
-  let totalCount = 0;
-  let products: Product[];
-  const conn = injector.get(TransactionalConnection);
-  await conn.startTransaction(ctx);
-  do {
-    // get all products of the source channel
-    products = await getProductsDeep(
-      ctx,
-      injector,
-      {
-        channels: { id: sourceChannelId },
-        deletedAt: IsNull(),
-      },
-      batch,
-      totalCount
-    );
-    totalCount += products.length;
-    // assign them to target channel
-    await assignTheseProductsToChannel(
-      targetChannelId,
-      injector,
-      products,
-      ctx
-    );
-  } while (products.length);
-  await conn.commitOpenTransaction(ctx);
+  await assignProductsInBatch(
+    targetChannelId,
+    {
+      channels: { id: sourceChannelId },
+      deletedAt: IsNull(),
+    },
+    injector,
+    ctx,
+    batch
+  );
 }
