+++ packages/vendure-scripts/src/assign-products-to-channel/assign-these-products-to-channel.ts
@@ -1,15 +1,12 @@
 import {
   AssetService,
-  ChannelService,
   FacetService,
   ID,
   Injector,
   Product,
   ProductVariantService,
   RequestContext,
-  SearchService,
 } from '@vendure/core';
-import { getSuperadminContextInChannel } from '../../../util/src/superadmin-request-context';
 
 /**
  * This function should be used with batches of Products
@@ -24,8 +21,6 @@ export async function assignTheseProductsToChannel(
   const productVariantService = injector.get(ProductVariantService);
   const assetService = injector.get(AssetService);
   const facetService = injector.get(FacetService);
-  const channelService = injector.get(ChannelService);
-  const searchService = injector.get(SearchService);
   const facetIds = products
     .map((product) =>
       (product.facetValues ?? []).map((facetValue) => facetValue.facet.id)
@@ -37,11 +32,6 @@ export async function assignTheseProductsToChannel(
   const variantIds = products
     .map((product) => (product.variants ?? []).map((variant) => variant.id))
     .flat();
-  const targetChannel = await channelService.findOne(ctx, targetChannelId);
-  const superadminContextInTargetChannel = await getSuperadminContextInChannel(
-    injector,
-    targetChannel!
-  );
   await Promise.all([
     //we will first assign the product variants, which will also assign the associated products and variants' assets
     productVariantService.assignProductVariantsToChannel(ctx, {
@@ -55,6 +45,5 @@ export async function assignTheseProductsToChannel(
       channelId: targetChannelId,
       facetIds,
     }),
-    searchService.reindex(superadminContextInTargetChannel),
   ]);
 }
